[RRTB-12] move all messaging classes to separate packageˆ
