+++ rrtb_input_tbot/src/main/java/com/nb/service/messaging/DlqMessageProducer.java
@@ -1,4 +1,4 @@
-package com.nb.service;
+package com.nb.service.messaging;
 
 import io.micronaut.jms.annotations.JMSProducer;
 import io.micronaut.jms.annotations.Queue;
